some changes

+++ internal/ports/presenters.go
@@ -15,7 +15,7 @@ type Message struct {
 	ID     int64  `json:"id"`
 	Date   int64  `json:"date"`
 	PeerID int64  `json:"peer_id"`
-	FromID int64  `json:"from_id"`
+	FromID int    `json:"from_id"`
 	Text   string `json:"text"`
 }
